Add Main as library
