+++ .github/workflows/go.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up Go 1.x
         uses: actions/setup-go@v2
         with:
-          go-version: ^1.16
+          go-version: ^1.17
         id: go
 
       - name: Check out code into the Go module directory
@@ -41,11 +41,5 @@ jobs:
       - name: Build
         run: go build -v .
 
-      - name: Unit Test
+      - name: Run Tests
         run: go test -v -race ./...
-
-      - name: Integration Test Unordered
-        run: go run cmd/sacura/main.go --config $(pwd)/test/config.yaml
-
-      - name: Integration Test Ordered
-        run: go run cmd/sacura/main.go --config $(pwd)/test/config-ordered.yaml
